feat(Aggregator): 로또 결과를 집계한다

+++ src/main/java/lotto/LottoFactory.java
@@ -10,11 +10,11 @@ public class LottoFactory {
     }
 
     public List<Lotto> createLottos(Money money) {
-        int amount = money.amount();
-        int count = amount / Money.UNIT;
+        long amount = money.amount();
+        long count = amount / Money.UNIT;
 
         List<Lotto> results = new ArrayList<>();
-        for (int i = 0; i < count; i++) {
+        for (long i = 0; i < count; i++) {
             List<Integer> randomUniqueNumbers = Randoms.pickUniqueNumbersInRange(1, 45, 6);
             results.add(Lotto.of(randomUniqueNumbers));
         }
